reordered quotient polynomial terms
we use field element 'alpha' to ensure that the identities in the
quotient polynomial can be considered linearly independent
from one another.

The 'arithmetic check' is specified as the first term in the quotient
polynomial, mulitplied by 'alpha', with 'alpha^2', 'alpha^3', 'alpha^4'
used to separate out the identities used to validate the permutation.

When adding additional gate types into PLONK, we have this odd
separation, where the arithmetic widget requires 'alpha' as its random
'separator' variable, and any extra arithmetic terms start with
'alpha^5'.

The goal of the polymorphic 'widget' construct is to encapsulate the
logic for a custom gate into a 'widget', and then pass a container of
widget pointers into the prover/verifier. Having the current widget
feed the random separator variable into the next widget,
when computing quotient polynomial terms, allows us to not
map specific powers of 'alpha' to specific gate types.

Long story short, 'alpha' to 'alpha^3' are reserved for permutation
terms in the quotient polynomial. The first 'widget' is fed 'alpha^4'
as its random separator variable - the next free 'alpha' power
is output by `widget::compute_quotient_contribution', which is
fed into the next widget, etc.