homogenized field operations
fq.hpp and fr.hpp are now typdef wrappers around a generic
`field` class, that implements all field operations

This work is part of a refactor to support multiple elliptic curves.
We need to instantiate an elliptic curve where the roles of fq and
fq are swapped around, and we want to avoid duplicating code.

Also, the current codebase had far too much code duplication
between fq and fr, given the similarities.