added Transcript class
the Transcript class is designed to simplify generating proof
transcripts, and the application of the Fiat-Shamir heuristic.

The Transcript class is agnostic as to the underlying proof system. A proof
is described via a ProgramManifest object.

The manifest specifies:

* the number of rounds of interaction
* the names and sizes of elements added to the transcript at each round
* whether transcript elements will be derived by the verifier or supplied
by the prover

Currently a work in progress, a couple of tests aren't passing yet