collision-resisant pedersen compression added
the pedersen compression function will take two field elements and
compute a pedersen hash of the result - returning a field element.

this can be used as a collision-resistant 'hash' function for
membersip / non-membership proofs in Merkle trees.

the method is called `compress` instead of `hash`, as Pedersen hashes
aren't suitable for when random oracles and pseudo random number
generators are required.

the compression function simultaneously computes a pedersen hash
of two native field elements, and an in-situ quaternary decomposition
of the input field elements.

this means that compression function outputs can be fed in as
compression function inputs, with zero overhead (i.e. what we need
for Merkle trees).

total number of gates for the compression function = 262. which is
~2 bits per gate (n.b. 2 bits per gate, *not* 2 gates per bit)