added faster version of pippenger
This new variant is ~45% faster than the old pippenger algorithm!

It uses affine addition formulae to compute group operations.
Previously, we used Jacobian coordinates to avoid having
to compute a modular inverse.

The new algorithm instead amortises the cost of the inverse away,
by using Montgomery's batch inversion technique.

This reduces the number of mulmods per group op from 11 to 6.

This is borne out in our benchmarks,
which seem to indicate a 45% speed-up.

Currently implemented in the `pippenger_unsafe` method. Needs a lot of
polish; the sorting algorithm is a bit ad-hoc right now.