added explicit template instantiation
most of our template usage is so that we
can create generic standard library types
and methods that support generic composers.

However, the set of all composers is small and fixed, so this is an
ideal use case for explicit template instantiation.

Most .tcc files have been converted into .cpp files, which seems to
have a positive effect on compile times.