refactored widgets into random widgets and transition widgets
The previous method of implementing custom gates resulted in code
duplication and did not allow the developer to succinctly express the
logic defined by a given custom gate.

widgets are now split into 'random' widgets and 'transition' widgets.
The former use randomness as part of their polynomial identity and
are generally more complex (e.g. the permutation widget and the
plookup widget).

Transition widgets define a series of identities that are evaluated when
transitioning between two rows in program memory. Generally the
lack of randomness within a gate makes these simpler to express.

Permutation widgets use the old widget structure due to how custom
the current random widgets are.

Transition widgets are created by instantiating the TransitionWidget class
with a Kernel template parameter.

The Kernel class has two key methods `compute_linear_terms` and
`compute_non_linear_terms`. These two methods define the component
of the polynomial identity specific to a widget that is included in the
linearisation trick, and any component that is not.

These two methods are used by TransitionWiget to construct:

1: The widget's contribution to the quotient polynomial
2: The widget's contribution to the linearisation polynomial
3: The widget's contribution to the verifier quotient polynomial reconstruction
4: The widget's contribution to the list of scalar multipliers used to
open the proof's batch of Kate commitments

Previously these 4 components had to be defined separately, resulting
in a lot of code duplication to express what is fundamentally the same
polynomial identity 4 times over.