fixed pippenger error
segfaults were triggering when the number of points was not a
clean power of two

we pre-allocate memory to hold the multiexponentiation
'point schedule', which describes which bucket to assign
a point to for a given round.

when the number of points is not a clean power of two,
we were sometimes underestimating the size of the
point schedule.

The error case was triggered because `get_num_rounds` has
breakpoints that are not powers of 2.

If `num_rounds(x) < num_rounds(floor_base_2(x))`, not enough
memory was being allocated to the point schedule - as we
implement pippenger by splitting the points into powers of two
and recursively calling `pippenger` to evalute the complete multiexp.
But all recursive calls used the same memory for their point schedules.