Zw/recursion variable keys (#53)
* removed unused members from proving key

* ComposerBase::compute_witness tweaks

the final witness element is now random - this ensures the witness commitments are never points at infinity

* recursive verifier tweaks

recursive verifier g1 exponentiation size
is no longer affected by whether proof commitments are valid

was causing problems, where the circuit created by feeding in junk witness data
was different to the circuit created with a satisfying witness

* added stdlib pow method

computes a**b where a is a `stdlib:field_t` and b is a `stdlib::uint32`

needed to recursion

* fixed stdlib::bigfield bug

`self_reduce` would segfault if called on
a constant

* added `from_witness` bigfield/biggroup methods

helper methods to convert a field/group element into a circuit witness
and return a bigfield/biggroup object

* stdlib::recursion::verifier accepts arbitrary verification keys

previously, the verification key was represented as a part of the
recursive verifier circuit
i.e. if circuit A verifies circuit B, circuit B's verification key
would be described via circuit A's
precomputed selector polynomials

The inner verification key can now be
represented either via constants or
witnesses. The latter will treat the key's
selector polynomials as private inputs to the outer verification circuit

this required the following changes:

* methods that accepted a verification key pointer now define the key
type via a template parameter

* added basic passing test for variable verification keys

* fixed variable verification key test bugs

* added `validate_key_is_in_set` method to recursive verification key

* added failure case test to recursive verifier

* updated rollup code to use new stdlib verification_key

* fixed bug where pedersen::compress would segfault when fed constants

* recursive verifier bug fixes

verifier will now succesfully validate proofs when fed a constant verification key

* fixed clang compiler error

* fixed stdlib::field_t division bug

the `/` operator was broken for cases where
the numerator was constant and the denominator
was a witness

* stdlib verification key tweaks

inverted values in the verification key are now validated instead of
being supplied as direct witnesses

Co-authored-by: Charlie Lye <karl.lye@gmail.com>